src/requirements.txt
bs4
click

gspread
google-api-python-client
pandas
requests
selenium
webdriver-manager